requirements.txt
# opencv-python==4.6.0.66

# pillow==9.2.0

# face_recognition==1.3.0


easydict
numpy
tqdm
torchvision
torch
opencv_python
Pillow
tensorboardX
